move MeatAxe into own file
